日志修改

+++ cloud-provider-payment8001/src/main/java/com/gguoliang/springcloud/mapper/PaymentMapper.java
@@ -23,6 +23,6 @@ public interface PaymentMapper {
      * @param paymentId
      * @return
      */
-    Payment getPaymentById(int paymentId);
+    Payment getPaymentById(Long paymentId);
 
 }
