+++ cloud-provider-payment8001/src/main/java/com/gguoliang/springcloud/service/PaymentService.java
@@ -22,5 +22,5 @@ public interface PaymentService {
      * @param paymentId
      * @return
      */
-    Payment getPaymentById(int paymentId);
+    Payment getPaymentById(Long paymentId);
 }
